+++ internal/checks/query_cost_test.go
@@ -9,6 +9,7 @@ import (
 	"time"
 
 	"github.com/cloudflare/pint/internal/checks"
+	"github.com/cloudflare/pint/internal/promapi"
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/rs/zerolog"
@@ -79,12 +80,12 @@ func TestCostCheck(t *testing.T) {
 		{
 			description: "ignores rules with syntax errors",
 			content:     "- record: foo\n  expr: sum(foo) without(\n",
-			checker:     checks.NewCostCheck("prom", "http://localhost", time.Second*5, 4096, 0, checks.Bug),
+			checker:     checks.NewCostCheck(promapi.NewPrometheus("prom", "http://localhost", time.Second*5), 4096, 0, checks.Bug),
 		},
 		{
 			description: "empty response",
 			content:     content,
-			checker:     checks.NewCostCheck("prom", srv.URL+"/empty/", time.Second*5, 4096, 0, checks.Bug),
+			checker:     checks.NewCostCheck(promapi.NewPrometheus("prom", srv.URL+"/empty/", time.Second*5), 4096, 0, checks.Bug),
 			problems: []checks.Problem{
 				{
 					Fragment: "sum(foo)",
@@ -98,7 +99,7 @@ func TestCostCheck(t *testing.T) {
 		{
 			description: "response timeout",
 			content:     content,
-			checker:     checks.NewCostCheck("prom", srv.URL+"/empty/", time.Millisecond*5, 4096, 0, checks.Bug),
+			checker:     checks.NewCostCheck(promapi.NewPrometheus("prom", srv.URL+"/empty/", time.Millisecond*5), 4096, 0, checks.Bug),
 			problems: []checks.Problem{
 				{
 					Fragment: "sum(foo)",
@@ -112,7 +113,7 @@ func TestCostCheck(t *testing.T) {
 		{
 			description: "bad request",
 			content:     content,
-			checker:     checks.NewCostCheck("prom", srv.URL+"/400/", time.Second*5, 4096, 0, checks.Bug),
+			checker:     checks.NewCostCheck(promapi.NewPrometheus("prom", srv.URL+"/400/", time.Second*5), 4096, 0, checks.Bug),
 			problems: []checks.Problem{
 				{
 					Fragment: "sum(foo)",
@@ -126,7 +127,7 @@ func TestCostCheck(t *testing.T) {
 		{
 			description: "1 result",
 			content:     content,
-			checker:     checks.NewCostCheck("prom", srv.URL+"/1/", time.Second*5, 4096, 0, checks.Bug),
+			checker:     checks.NewCostCheck(promapi.NewPrometheus("prom", srv.URL+"/1/", time.Second*5), 4096, 0, checks.Bug),
 			problems: []checks.Problem{
 				{
 					Fragment: "sum(foo)",
@@ -140,7 +141,7 @@ func TestCostCheck(t *testing.T) {
 		{
 			description: "7 results",
 			content:     content,
-			checker:     checks.NewCostCheck("prom", srv.URL+"/7/", time.Second*5, 101, 0, checks.Bug),
+			checker:     checks.NewCostCheck(promapi.NewPrometheus("prom", srv.URL+"/7/", time.Second*5), 101, 0, checks.Bug),
 			problems: []checks.Problem{
 				{
 					Fragment: "sum(foo)",
@@ -154,7 +155,7 @@ func TestCostCheck(t *testing.T) {
 		{
 			description: "7 result with MB",
 			content:     content,
-			checker:     checks.NewCostCheck("prom", srv.URL+"/7/", time.Second*5, 1024*1024, 0, checks.Bug),
+			checker:     checks.NewCostCheck(promapi.NewPrometheus("prom", srv.URL+"/7/", time.Second*5), 1024*1024, 0, checks.Bug),
 			problems: []checks.Problem{
 				{
 					Fragment: "sum(foo)",
@@ -168,7 +169,7 @@ func TestCostCheck(t *testing.T) {
 		{
 			description: "7 results with 1 series max (1KB bps)",
 			content:     content,
-			checker:     checks.NewCostCheck("prom", srv.URL+"/7/", time.Second*5, 1024, 1, checks.Bug),
+			checker:     checks.NewCostCheck(promapi.NewPrometheus("prom", srv.URL+"/7/", time.Second*5), 1024, 1, checks.Bug),
 			problems: []checks.Problem{
 				{
 					Fragment: "sum(foo)",
@@ -182,7 +183,7 @@ func TestCostCheck(t *testing.T) {
 		{
 			description: "7 results with 5 series max",
 			content:     content,
-			checker:     checks.NewCostCheck("prom", srv.URL+"/7/", time.Second*5, 0, 5, checks.Bug),
+			checker:     checks.NewCostCheck(promapi.NewPrometheus("prom", srv.URL+"/7/", time.Second*5), 0, 5, checks.Bug),
 			problems: []checks.Problem{
 				{
 					Fragment: "sum(foo)",
@@ -196,7 +197,7 @@ func TestCostCheck(t *testing.T) {
 		{
 			description: "7 results with 5 series max / infi",
 			content:     content,
-			checker:     checks.NewCostCheck("prom", srv.URL+"/7/", time.Second*5, 0, 5, checks.Information),
+			checker:     checks.NewCostCheck(promapi.NewPrometheus("prom", srv.URL+"/7/", time.Second*5), 0, 5, checks.Information),
 			problems: []checks.Problem{
 				{
 					Fragment: "sum(foo)",
@@ -213,7 +214,7 @@ func TestCostCheck(t *testing.T) {
 - record: foo
   expr: 'sum({__name__="foo"})'
 `,
-			checker: checks.NewCostCheck("prom", srv.URL+"/7/", time.Second*5, 101, 0, checks.Bug),
+			checker: checks.NewCostCheck(promapi.NewPrometheus("prom", srv.URL+"/7/", time.Second*5), 101, 0, checks.Bug),
 			problems: []checks.Problem{
 				{
 					Fragment: `sum({__name__="foo"})`,