+++ internal/config/rule.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/cloudflare/pint/internal/checks"
 	"github.com/cloudflare/pint/internal/parser"
+	"github.com/cloudflare/pint/internal/promapi"
 	"github.com/rs/zerolog/log"
 )
 
@@ -155,8 +156,8 @@ type Rule struct {
 	Reject     []RejectSettings     `hcl:"reject,block" json:"reject,omitempty"`
 }
 
-func (rule Rule) resolveChecks(path string, r parser.Rule, enabledChecks, disabledChecks []string, proms []PrometheusConfig) []checks.RuleChecker {
-	enabled := []checks.RuleChecker{}
+func (rule Rule) resolveChecks(path string, r parser.Rule, enabledChecks, disabledChecks []string, prometheusServers []*promapi.Prometheus) []checkMeta {
+	enabled := []checkMeta{}
 
 	if rule.Ignore != nil && rule.Ignore.IsMatch(path, r) {
 		return enabled
@@ -174,69 +175,59 @@ func (rule Rule) resolveChecks(path string, r parser.Rule, enabledChecks, disabl
 			}
 			severity := aggr.getSeverity(checks.Warning)
 			for _, label := range aggr.Keep {
-				if isEnabled(enabledChecks, disabledChecks, checks.AggregationCheckName, r) {
-					enabled = append(enabled, checks.NewAggregationCheck(nameRegex, label, true, severity))
-				}
+				enabled = append(enabled, checkMeta{
+					name:  checks.AggregationCheckName,
+					check: checks.NewAggregationCheck(nameRegex, label, true, severity),
+				})
 			}
 			for _, label := range aggr.Strip {
-				if isEnabled(enabledChecks, disabledChecks, checks.AggregationCheckName, r) {
-					enabled = append(enabled, checks.NewAggregationCheck(nameRegex, label, false, severity))
-				}
+				enabled = append(enabled, checkMeta{
+					name:  checks.AggregationCheckName,
+					check: checks.NewAggregationCheck(nameRegex, label, false, severity),
+				})
 			}
 		}
 	}
 
-	if isEnabled(enabledChecks, disabledChecks, checks.RateCheckName, r) {
-		for _, prom := range proms {
-			timeout, _ := parseDuration(prom.Timeout)
-			enabled = append(enabled, checks.NewRateCheck(prom.Name, prom.URI, timeout))
-		}
-	}
-
-	if isEnabled(enabledChecks, disabledChecks, checks.SeriesCheckName, r) {
-		for _, prom := range proms {
-			timeout, _ := parseDuration(prom.Timeout)
-			enabled = append(enabled, checks.NewSeriesCheck(prom.Name, prom.URI, timeout))
-		}
-	}
-
-	if isEnabled(enabledChecks, disabledChecks, checks.VectorMatchingCheckName, r) {
-		for _, prom := range proms {
-			timeout, _ := parseDuration(prom.Timeout)
-			enabled = append(enabled, checks.NewVectorMatchingCheck(prom.Name, prom.URI, timeout))
-		}
-	}
-
-	if rule.Cost != nil && isEnabled(enabledChecks, disabledChecks, checks.CostCheckName, r) {
+	if rule.Cost != nil {
 		severity := rule.Cost.getSeverity(checks.Bug)
-		for _, prom := range proms {
-			timeout, _ := parseDuration(prom.Timeout)
-			enabled = append(enabled, checks.NewCostCheck(prom.Name, prom.URI, timeout, rule.Cost.BytesPerSample, rule.Cost.MaxSeries, severity))
+		for _, prom := range prometheusServers {
+			enabled = append(enabled, checkMeta{
+				name:  checks.CostCheckName,
+				check: checks.NewCostCheck(prom, rule.Cost.BytesPerSample, rule.Cost.MaxSeries, severity),
+			})
 		}
 	}
 
-	if len(rule.Annotation) > 0 && isEnabled(enabledChecks, disabledChecks, checks.AnnotationCheckName, r) {
+	if len(rule.Annotation) > 0 {
 		for _, ann := range rule.Annotation {
 			var valueRegex *regexp.Regexp
 			if ann.Value != "" {
 				valueRegex = strictRegex(ann.Value)
 			}
 			severity := ann.getSeverity(checks.Warning)
-			enabled = append(enabled, checks.NewAnnotationCheck(ann.Key, valueRegex, ann.Required, severity))
+			enabled = append(enabled, checkMeta{
+				name:  checks.AnnotationCheckName,
+				check: checks.NewAnnotationCheck(ann.Key, valueRegex, ann.Required, severity),
+			})
 		}
 	}
-	if len(rule.Label) > 0 && isEnabled(enabledChecks, disabledChecks, checks.LabelCheckName, r) {
+
+	if len(rule.Label) > 0 {
 		for _, lab := range rule.Label {
 			var valueRegex *regexp.Regexp
 			if lab.Value != "" {
 				valueRegex = strictRegex(lab.Value)
 			}
 			severity := lab.getSeverity(checks.Warning)
-			enabled = append(enabled, checks.NewLabelCheck(lab.Key, valueRegex, lab.Required, severity))
+			enabled = append(enabled, checkMeta{
+				name:  checks.LabelCheckName,
+				check: checks.NewLabelCheck(lab.Key, valueRegex, lab.Required, severity),
+			})
 		}
 	}
 
-	if rule.Alerts != nil && isEnabled(enabledChecks, disabledChecks, checks.AlertsCheckName, r) {
+	if rule.Alerts != nil {
 		qRange := time.Hour * 24
 		if rule.Alerts.Range != "" {
 			qRange, _ = parseDuration(rule.Alerts.Range)
@@ -249,30 +240,41 @@ func (rule Rule) resolveChecks(path string, r parser.Rule, enabledChecks, disabl
 		if rule.Alerts.Resolve != "" {
 			qResolve, _ = parseDuration(rule.Alerts.Resolve)
 		}
-		for _, prom := range proms {
-			timeout, _ := parseDuration(prom.Timeout)
-			enabled = append(enabled, checks.NewAlertsCheck(prom.Name, prom.URI, timeout, qRange, qStep, qResolve))
+		for _, prom := range prometheusServers {
+			enabled = append(enabled, checkMeta{
+				name:  checks.AlertsCheckName,
+				check: checks.NewAlertsCheck(prom, qRange, qStep, qResolve),
+			})
 		}
 	}
 
-	if len(rule.Reject) > 0 && isEnabled(enabledChecks, disabledChecks, checks.RejectCheckName, r) {
+	if len(rule.Reject) > 0 {
 		for _, reject := range rule.Reject {
 			severity := reject.getSeverity(checks.Bug)
+			re := strictRegex(reject.Regex)
 			if reject.LabelKeys {
-				re := strictRegex(reject.Regex)
-				enabled = append(enabled, checks.NewRejectCheck(true, false, re, nil, severity))
+				enabled = append(enabled, checkMeta{
+					name:  checks.RejectCheckName,
+					check: checks.NewRejectCheck(true, false, re, nil, severity),
+				})
 			}
 			if reject.LabelValues {
-				re := strictRegex(reject.Regex)
-				enabled = append(enabled, checks.NewRejectCheck(true, false, nil, re, severity))
+				enabled = append(enabled, checkMeta{
+					name:  checks.RejectCheckName,
+					check: checks.NewRejectCheck(true, false, nil, re, severity),
+				})
 			}
 			if reject.AnnotationKeys {
-				re := strictRegex(reject.Regex)
-				enabled = append(enabled, checks.NewRejectCheck(false, true, re, nil, severity))
+				enabled = append(enabled, checkMeta{
+					name:  checks.RejectCheckName,
+					check: checks.NewRejectCheck(false, true, re, nil, severity),
+				})
 			}
 			if reject.AnnotationValues {
-				re := strictRegex(reject.Regex)
-				enabled = append(enabled, checks.NewRejectCheck(false, true, nil, re, severity))
+				enabled = append(enabled, checkMeta{
+					name:  checks.RejectCheckName,
+					check: checks.NewRejectCheck(false, true, nil, re, severity),
+				})
 			}
 		}
 	}
@@ -280,16 +282,24 @@ func (rule Rule) resolveChecks(path string, r parser.Rule, enabledChecks, disabl
 	return enabled
 }
 
-func isEnabled(enabledChecks, disabledChecks []string, name string, rule parser.Rule) bool {
-	if rule.HasComment(fmt.Sprintf("disable %s", removeRedundantSpaces(name))) {
-		log.Debug().
-			Str("check", name).
-			Msg("Check disabled by comment")
-		return false
+func isEnabled(enabledChecks, disabledChecks []string, rule parser.Rule, name string, check checks.RuleChecker) bool {
+	instance := check.String()
+	comments := []string{
+		fmt.Sprintf("disable %s", name),
+		fmt.Sprintf("disable %s", instance),
+	}
+	for _, comment := range comments {
+		if rule.HasComment(comment) {
+			log.Debug().
+				Str("check", instance).
+				Str("comment", comment).
+				Msg("Check disabled by comment")
+			return false
+		}
 	}
 
 	for _, c := range disabledChecks {
-		if c == name {
+		if c == name || c == instance {
 			return false
 		}
 	}